+++ src/Behat/Behat/Output/ServiceContainer/Formatter/PrettyFormatterFactory.php
@@ -167,7 +167,7 @@ class PrettyFormatterFactory implements FormatterFactory
      */
     protected function loadFormatter(ContainerBuilder $container)
     {
-        $definition = new Definition('Behat\Behat\Output\Statistics\Statistics');
+        $definition = new Definition('Behat\Behat\Output\Statistics\TotalStatistics');
         $container->setDefinition('output.pretty.statistics', $definition);
 
         $definition = new Definition('Behat\Testwork\Output\NodeEventListeningFormatter', array(
