+++ src/Behat/Behat/Output/ServiceContainer/Formatter/ProgressFormatterFactory.php
@@ -135,7 +135,7 @@ class ProgressFormatterFactory implements FormatterFactory
      */
     protected function loadFormatter(ContainerBuilder $container)
     {
-        $definition = new Definition('Behat\Behat\Output\Statistics\Statistics');
+        $definition = new Definition('Behat\Behat\Output\Statistics\TotalStatistics');
         $container->setDefinition('output.progress.statistics', $definition);
 
         $definition = new Definition('Behat\Testwork\Output\NodeEventListeningFormatter', array(
